update error JSON schemas (both schemas)
represent "position" as a tree, fixing problems raised in meeting 009. Also, bear in mind that now "None" is allowed as a value only for the key 'item', since it is not possible to access the items inside a completely empty string (cell). However, if the string contains anything (including spaces only), items' indexes should be used instead, to represent what is wrong inside the cell(s).