begin structure in classes
N.B.: FROM NOW ON, CHANGES/ADDITIONS MUST BE MADE ON THE FILES INSIDE THE "validator" DIRECTORY!! CHANGES OUTSIDE OF IT MIGHT **ALSO** BE MADE, BUT THEY WILL BE FOR ARCHIVAL PURPOSES ONLY!
many additions in the main functions, but only on the files in the "validator" folder. I have begun structuring the code in classes. Also, added in validate_cits and validate_meta the calls for the function for lev2 and lev3. Function for lev4 is yet to be implemented, but the id_types_alignment is ready ("id_types_alignment.json" inside "validator" directory).